pyproject.toml
[tool.poetry]
name = "nerf-qa"
version = "0.3.0"
description = ""
authors = ["kobejean <s1292010@u-aizu.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
opencv_python = "^4.9.0"
pandas = "^2.2.0"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"
tqdm = "^4.66.2"
argparse = "^1.4.0"
wandb = "^0.16.3"
dists-pytorch = "^0.1"
plotly = "^5.19.0"
nbformat = "^5.9.2"
pyfeats = "^1.0.1"
einops = "^0.7.0"
kornia = "^0.7.1"
albumentations = "^1.4.0"
loguru = "^0.7.2"
matplotlib = "^3.8.3"
h5py = "^3.10.0"
timm = "^0.9.16"
xformers = "^0.0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
